Handles video sources with codecs correctly
When using H264 and H264 they can both be provided within a mp4 container. To tell the browser which one to prefer we provide them in a specific order and pass in the codec as type. This resulted in an invalid selector.
We now instead keep a list of all created source elements during initialization and switch the source files that way.

Fixes: OX-6937