Displays a short overlay when a keyboard shortcut is triggered
This is done to signal that the keyboard shortcut registered correctly.
It also tells the user the new volume when altering it via the arrow keys.

The following actions currently result in an overlay:
- Spacebar to pause video
- Spacebar to play video
- Arrow up to increase volume
- Arrow down to decrease volume

This can later be expanded to display overlays on even more keyboard shortcuts but first more icons are needed.

Fixes: OX-6936