Allows to fall back to best available quality for playlist videos
Example:
- First video of the playlist supports 1080p and 1080p is selected
- Second video of the playlist only supports 720p

Previously this would fail as no 1080p source could be found when switching from video 1 to 2. We now fall back to the highest available quality for video 2 when switching to it.

This seems like the most graceful way to handle different video resolutions in the same playlist for now.

Fixes: OX-6936