Allows to provide different posters for each video of playlist
Also changes the sources parameter to a single video parameter which is turned into a playlist internally. This allows to have simplified code which always assumes a playlist but easier constructor options structure. Also the different video specific options will be moved into the video class so they are bundled together.

Fixes: OX-6936