Filters out quality levels without playable codec
When instantiating the player one can pass one or multiple videos with one or multiple quality levels. For each quality level one or multiple different codecs can be provided.
Previously the player did not check whether those codecs can actually be played by the browser, which resulted in quality levels that were selectable but did not result in a playable video.

Fixes: OX-10520