Hide included icon sheet container properly
The video player loads an SVG icon sheet and adds it to the bottom of the page. This container was previously not hidden and some browsers rendered it with a height higher than 0.
We now hide it to ensure its invisible and does not mess with the height of the page.